#54 カスタムフックに処理を切り出した
