+++ src/features/chat/components/ChatMainContainer/index.tsx
@@ -2,62 +2,21 @@ import { useRouter } from 'next/router';
 import { useCallback, useEffect, useState } from 'react';
 import { useSendChat } from '../../hooks/useSendChat';
 import type { SendChatArgument } from '../../hooks/useSendChat';
+import { useThreadId } from '../../hooks/useThreadId';
 import { useThreadMessages } from '../../hooks/useThreadMessages';
 import { ChatMain } from '../ChatMain';
 
 export type HandleSendChatType = (arg: SendChatArgument) => Promise<void>;
 
 export const ChatMainContainer = () => {
-  const [threadId, setThreadId] = useState<string | undefined>(undefined);
-
   const { sendChat, latestQuestion, latestAnswer, clearLatestAnswer } =
     useSendChat();
 
-  const router = useRouter();
-
-  /**
-   * ページを開いた際にURLパラメータからthreadIdを取得します。
-   */
-  useEffect(() => {
-    if (router.isReady) {
-      const id = getThreadId();
-      setThreadId(id);
-    }
-  }, [router.isReady]);
-
-  /**
-   * ルータークエリからthreadIdを取得します。
-   * @returns {string | undefined} 利用可能な場合、threadIdを返します。
-   */
-  const getThreadId = useCallback((): string | undefined => {
-    const queryThreadId = router.query.threadId;
-
-    if (Array.isArray(queryThreadId)) {
-      return queryThreadId[0];
-    }
-    return queryThreadId;
-  }, [router.query.threadId]);
+  const { threadId, setThreadId, getThreadId } = useThreadId();
 
   const { fetchThreadMessages } = useThreadMessages();
   const { data, refetch } = fetchThreadMessages(threadId);
   const chatLog = data ?? [];
-
-  /**
-   * 現在のthreadIdでURLパラメータの値を更新する。
-   */
-  useEffect(() => {
-    const currentThreadId = getThreadId();
-    if (currentThreadId && threadId && currentThreadId !== threadId) {
-      router.replace(
-        {
-          pathname: router.pathname,
-          query: { ...router.query, threadId },
-        },
-        undefined,
-        { shallow: true },
-      );
-    }
-  }, [getThreadId, threadId, router.pathname, router.replace, router.query]);
 
   /**
    * チャットメッセージを送信します。
